requirements-dev.txt
-r requirements.txt
pytest-django==3.1.2
pytest-cov==2.4.0
model-mommy==1.3.2
ipdb==0.10.2
